requirements.txt
firebase-admin
google-cloud-firestore
fastapi~=0.63.0
uvicorn~=0.13.4
python-dotenv~=0.17.0
pygame~=2.0.1
python-multipart
starlette~=0.13.6